+++ bin/src/commands/generate_command.dart
@@ -2,7 +2,7 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:args/command_runner.dart';
-import 'package:orm/configure.dart';
+import 'package:orm/configure.dart' as configure;
 import 'package:orm/dmmf.dart';
 import 'package:orm/generator_helper.dart';
 import 'package:orm/orm.dart';
@@ -14,7 +14,6 @@ import '../binary_engine/binray_engine_type.dart';
 import '../generator/generator.dart';
 import '../generator/generator_options.dart';
 import '../utils/ansi_progress.dart';
-import '../utils/find_project.dart';
 
 class GenerateCommand extends Command {
   @override
@@ -28,8 +27,7 @@ class GenerateCommand extends Command {
       'schema',
       help: 'Custom path to your Prisma schema',
       valueHelp: 'path',
-      defaultsTo:
-          configure('schema', joinRelativePaths(['prisma', 'schema.prisma'])),
+      defaultsTo: configure.schema,
     );
   }