+++ lib/src/configure/configure_web.dart
@@ -1,9 +1,0 @@
-import 'configure.dart';
-
-class _Web$Configure extends Configure {
-  @override
-  Map<String, dynamic> get all => const {};
-}
-
-/// Create web configure instance.
-Configure get configure => _Web$Configure();