+++ lib/src/configure/query_engine_type.dart
@@ -1,0 +1,5 @@
+/// Prisma query engine type.
+enum QueryEngineType {
+  binary,
+  library,
+}